Remove Shared With column from SelectApplication list, minor fixes
Remove Shared With column from SelectApplication list since it isn't necessary to display that data there.
Add Status column header to SelectApplication list for accessibility.
Add padding to enclosing box of clipart on SelectApplication screen to account for translation applied to clipart, so clipart does not exceed the bounds of the enclosing element.